Local account store delegate methods; methods for getting device serial number and mac address; category icons properties refactored; Moidom module dependencies injection fixed including TV directory providing; TV channel category entity; TV channel repository fixed; channel category entity mapper; channel entity mapper; channel local store delegate complate

+++ Domain/src/main/java/org/alsi/android/domain/tv/model/guide/TvChannel.kt
@@ -2,4 +2,4 @@ package org.alsi.android.domain.tv.model.guide
 
 import java.net.URI
 
-class TvChannel (val id: Long, val categoryId: Long, val logoUri: URI, val number: Long, val title: String)
+class TvChannel (val id: Long, val categoryId: Long, val logoUri: URI, val number: Int, val title: String)
